Declutter root: move setup scripts to scripts/, drop cached embeddings and sonar lock
- Move setup.sh, setup.bat, chrome_setup.sh, chrome_setup.bat and
  auto_install_detectron2.py into scripts/; update all references in
  CI, Dockerfile, README and the scripts themselves
- Remove tracked 30MB LOGO_FEATS.npy cache (already gitignored; regenerated on first run)
- Remove leftover .sonar/.sonar_lock
- Narrow overly-broad chrome* gitignore pattern to the actual download artifacts
- Translate remaining Chinese usage text in chrome_setup.sh
